Rename to Header and HeaderCell
